server/requirements.txt
tornado
pymongo
pytz
Pillow
pytesseract
itsdangerous
requests
unidecode